+++ app/src/main/java/com/github/kwasow/archipelago/views/GraphView.kt
@@ -195,9 +195,9 @@ class GraphView : View {
                 // Space data evenly
                 val date = (i + 1).toDouble()
                 val amount = if (i == 0) {
-                    transactions[0].amount
+                    transactions[0].amount.toDouble()
                 } else {
-                    returnList[i - 1][1] + transactions[i].amount
+                    returnList[i - 1][1] + transactions[i].amount.toDouble()
                 }
 
                 returnList.add(
